requirements.txt
pip install -r requirements.txt
# Core CDK libraries
aws-cdk-lib==2.134.0
constructs>=10.0.0,<11.0.0

# For testing CDK stacks
pytest
# AWS CDK Core Libraries
aws-cdk-lib==2.137.0
constructs>=10.0.0

# PySpark (for local testing of Glue scripts)
pyspark==3.4.1

# AWS SDK
boto3==1.34.89
botocore==1.34.89

# Unit Testing
pytest==8.2.0
pytest-mock==3.14.0

# Type Hints and Static Checks
typeguard==2.13.3

# Optional: YAML & JSON Handling (if using YAML-based schemas)
pyyaml==6.0.1

# Optional: AWS CLI (if using CLI from scripts)
awscli==1.32.89
pytest==6.2.5
pip install aws-cdk-lib constructs